fix: harden legacy organize copy paths
